replace funtions to utils
