+++ openfl/utilities/utils.py
@@ -3,8 +3,10 @@
 """Utilities module."""
 
 import hashlib
+import ipaddress
 import logging
 import os
+import re
 from socket import getfqdn
 
 import numpy as np
@@ -24,6 +26,36 @@ def getfqdn_env(name: str = '') -> str:
     if fqdn is not None:
         return fqdn
     return getfqdn(name)
+
+
+def is_fqdn(hostname: str) -> bool:
+    """https://en.m.wikipedia.org/wiki/Fully_qualified_domain_name."""
+    if not 1 < len(hostname) < 253:
+        return False
+
+    # Remove trailing dot
+    hostname.rstrip('.')
+
+    #  Split hostname into list of DNS labels
+    labels = hostname.split('.')
+
+    #  Define pattern of DNS label
+    #  Can begin and end with a number or letter only
+    #  Can contain hyphens, a-z, A-Z, 0-9
+    #  1 - 63 chars allowed
+    fqdn = re.compile(r'^[a-z0-9]([a-z-0-9-]{0,61}[a-z0-9])?$', re.IGNORECASE) # noqa 
+
+    # Check that all labels match that pattern.
+    return all(fqdn.match(label) for label in labels)
+
+
+def is_api_adress(address: str) -> bool:
+    """Validate ip-adress value."""
+    try:
+        ipaddress.ip_address(address)
+        return True
+    except ValueError:
+        return False
 
 
 def add_log_level(level_name, level_num, method_name=None):
